delete path analysis
path analysis methods are being deferred to a future version. SPLAT! does not output the actual table of height or rssi values as a function of distance between the transmitter and receiver.

Because of this, a good path analysis method requires reading the .hgt.gz and computing fresnel zone, etc. in python.

This is outside the scope of this project, which is just to provide a wrapper for the core SPLAT! functionality.

I may revisit this issue in the future.