add function to display and switch colorbars
frontend: added a set of png for the matplotlib colormaps we're using. Added a spot in the display section for the colormap and min/max rssi labels.

backend: added a utility to export matplotlib colormaps to png. if we want to add colormaps in the future, we run this and then copy the png to the static assets / colormaps directory.