add leaflet plugin to download the map, removed kml export and clear display buttons
We can't export the KML / KMZ easily since we implemented multiple sites (which is a more important feature anyway). For now I've added the leafelet "easyprint" button so the user can get a PNG of the map view with all their sites.

Later, we need to get rid of georasterlayer and switch to geojson, and at that point, KML output of all the sites will be possible.