SRTM v2 missing terrain tiles - fallback to v1 when needed.
the bucket v2/skadi contains NASA SRTM V2 tiles (see https://cmr.earthdata.nasa.gov/search/concepts/C1220566612-USGS_LTA.html). The V1 tiles are in prefix skadi/ which work fine, but the v2 tiles are not complete. a handful are missing which caused the backend to fail requests when it tried to download them.

this bugfix falls back to the v1 data when the more refined SRTM V2 data are not available in the AWS bucket.