bugfix - australia and europe coordinate error, backend failure
Fixed bug where the backend fails to download and convert the terrain tiles using srtm2sdf for coordinates in europe or australia. root cause was an option (-d) in srtm2sdf we don't need to call and an off-by-one error in the function which decides which tiles we need to pull from AWS.

tested in europe, north america, south america, and australia.